fix(empire-repository): Steve count() refactor + Norman test split
Steve / Norman PR #29 review feedback. 機能変更なし、品質改善のみ。

## Steve 指摘①: count() を proper SELECT COUNT(*) に差し替え

`SqliteEmpireRepository.count()` が `SELECT empires.id` を発行して
全行を Python に load してから `len()` を取っていた。設計書 §確定 D
の「`SELECT COUNT(*) FROM empires` の結果」契約に違反。

修正: `select(func.count()).select_from(EmpireRow)` で SQL レベル
の COUNT(*) を発行。Empire は singleton 前提だが、本 PR は §確定 F
で **後続 6 件 Repository PR の真実源** として凍結された
テンプレート責務を持つ。Workflow stages / Task deliverables 等で
全行 load パターンが伝播すると N+1 / メモリ無駄ロードが量産される
ため、最初の 1 行で正しい SQL を出すことが重要。

docstring に「テンプレート責務として COUNT(*) を発行する」根拠も
明記、後続 PR が真似していい状態に。

## Norman 指摘①: test_empire_repository.py を 506 行 → 3 ファイル分割

500 行ルール違反 + 後続 6 件 × 500+ 行が積み上がる退行リスク。
Norman の提案構造で `test_empire_repository/` ディレクトリに分割:

| 新ファイル | 行数 | 含む TC |
|---|---|---|
| `__init__.py` | 15 | 分割理由 docstring |
| `test_protocol_crud.py` | 171 | TC-IT-EMR-001/002/003/004/005/010/018 (Protocol + 基本 CRUD) |
| `test_save_semantics.py` | 236 | TC-IT-EMR-006/007/011/012 + TC-UT-EMR-003 (delete-then-insert + Tx 境界 + round-trip) |
| `test_constraints_arch.py` | 147 | TC-IT-EMR-013/014/017 (FK CASCADE + UNIQUE + CI 三層防衛 template) |

各ファイル 147〜236 行で readability budget 内、後続 Repository PR
が同じトピック分割を真似できる構造。テスト本体は変更なし、
import / class 配置のみ移動。

## Quality gates (local)

- pyright --strict: 0 errors / 0 warnings
- ruff check / format: clean
- pytest: **554 passed** (分割前と同数、無回帰)

Refs: #25, Steve / Norman PR #29 review
